# Migrate `glu` from the TH to Aten (CUDA)