--- a/aten/src/THCUNN/GatedLinearUnit.cu
+++ b/aten/src/THCUNN/GatedLinearUnit.cu
@@ -5,18 +5,6 @@
 #include <THCUNN/common.h>
 #include <ATen/WrapDimUtils.h>
 
-template <typename Dtype, typename Acctype>
-struct gatedLinearCSigMul_functor
-{
-  __device__ void operator()(Dtype *target, const Dtype *sigTensor, const Dtype *mulTensor) const
-  {
-    const Acctype sigNum = Acctype(1)/(Acctype(1)+ exp(ScalarConvert<Dtype, Acctype>::to(-*sigTensor)));
-    const Dtype mulNum = *mulTensor;
-    *target = ScalarConvert<Acctype, Dtype>::to(sigNum * mulNum);
-  }
-};
-
-
 template<typename Dtype, typename Acctype>
 struct gatedLinearDerivative
 {