Fix #24571

I tried to implement `glu_backward` in #39586 but encountered some performance issues, so this PR is opened only for the forward part. I'll clean up THCUNN in the PR for backward. 

# Benchmark

Built with `python setup.py develop`, running on a Tesla P100 GPU.

- Test script (modified from https://github.com/pytorch/pytorch/pull/33179):
```py
import torch
import torch.nn.functional as F
import time

torch.manual_seed(0)

def _time():
    if torch.cuda.is_available():
        torch.cuda.synchronize()
    return time.time()

device = "cuda"

#warm up
for n in [10, 100, 1000, 10000]:
    input = torch.randn(128, n, requires_grad=True, device=device)
    grad_output = torch.ones(128, n // 2, device=device)
    for i in range(1000):
        output = F.glu(input)
        output.backward(grad_output)

for n in [10, 100, 1000, 10000]:
    fwd_t = 0
    bwd_t = 0
    input = torch.randn(128, n, requires_grad=True, device=device)
    grad_output = torch.ones(128, n // 2, device=device)
    for i in range(10000):
        t1 = _time()
        output = F.glu(input)
        t2 = _time()
        output.backward(grad_output)
        t3 = _time()
        fwd_t = fwd_t + (t2 -t1)
        bwd_t = bwd_t + (t3 - t2)
    fwd_avg = fwd_t / 10000 * 1000
    bwd_avg = bwd_t / 10000 * 1000
    print("input size(128, %d) forward time is %.2f (ms); backwad avg time is %.2f (ms)."
          % (n, fwd_avg, bwd_avg))
```

- Before (as of commit [`9bfb91b50b53d332493d5179d3e17871f92efd2a`](https://github.com/pytorch/pytorch/tree/9bfb91b50b53d332493d5179d3e17871f92efd2a))
```
input size(128, 10) forward time is 0.08 (ms); backwad avg time is 0.17 (ms).
input size(128, 100) forward time is 0.08 (ms); backwad avg time is 0.17 (ms).
input size(128, 1000) forward time is 0.07 (ms); backwad avg time is 0.17 (ms).
input size(128, 10000) forward time is 0.07 (ms); backwad avg time is 0.17 (ms).
```

- After
```
input size(128, 10) forward time is 0.08 (ms); backwad avg time is 0.17 (ms).
input size(128, 100) forward time is 0.08 (ms); backwad avg time is 0.17 (ms).
input size(128, 1000) forward time is 0.08 (ms); backwad avg time is 0.17 (ms).
input size(128, 10000) forward time is 0.08 (ms); backwad avg time is 0.17 (ms).
```

cc: @VitalyFedyunin